Adjust the framework, Abstract the whole process of seckill from entering to getting goods into a call chain, and use the filter chain to complete the whole call. The data of each thread is saved to ThreadLocal.
